api/requirements.txt
torch>=2.0.0
diffusers>=0.24.0
transformers>=4.30.0
fastapi>=0.104.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pillow>=10.0.0
pydantic>=2.4.0
numpy>=1.24.0
accelerate>=0.25.0
safetensors>=0.4.0
python-dotenv>=1.0.0